Fixed FXAA! Now using the actual algorithm instead of simplified version.

* Fixed a bug where using unsigned values resulted in value wrapping.
* Many FXAA algorithm changes, including using 8 neighbouring (all) pixels instead of cross-pattern for better AA.

FXAA edge detection algorithm modified!
SSAA Implemented!

* SSAA member added to Antialiasing enum.
* EdgeDetection enum class added.
* Camera::GenerateRay() signature changed. Row and Col are now floats instead of ints.
* ColorUtils ges a new getColorDifference() function, calculating Euclidean distance.
* New toFloat inline function added to Render.cpp for quick and short unsigned->float casting.
* Render::RenderImage() and ::RnederTree() algorithm changed to account for SSAA.
* Settings:: subPixDepthAA and edgeDetectionTypeFXAA members added.
* ray2NDC space conversion function signature changed to take row and column as floats instead of ints.
* ray2NDC no longer multiplies by 0.5. This logic was moved to the caller.
* FVector3::CalculateLength() now uses sqrtf instead of sqrt.

+++ RayTracer/inc/SpaceConversions.h
@@ -10,7 +10,7 @@ struct ImagePlane;
 // @param[in] y: The y location of the center of the pixel.
 // @param[in] imgPlane: The camera's image plane used to get render resolution.
 // @return A vector representing the NDC coordinates in [0.0, 1.0] range.
-FVector2 ray2NDC( const int, const int, const ImagePlane& );
+FVector2 ray2NDC( const float, const float, const ImagePlane& );
 
 // Convert NDC to Screen Space (NSC to NDC - Normalized Device Coordinates) [-1.0, 1.0]
 // @param[in] coords: A vector with the NDC coordinates in [0.0, 1.0] range.
