+++ RayTracer/src/Vectors.cpp
@@ -1,6 +1,6 @@
 #include "Vectors.h"
 
-#include <cmath> // sqrt
+#include <cmath> // sqrtf
 #include <stdexcept> // out_of_range
 
 
@@ -114,7 +114,7 @@ void FVector3::NormalizeInPlace() {
 #pragma warning(disable : 4756)
 
 void FVector3::CalculateLength() {
-	length = std::sqrt( x * x + y * y + z * z );
+	length = std::sqrtf( x * x + y * y + z * z );
 }
 
 #pragma warning(pop)
